Users can now edit their own accounts and view others accounts. Users can also Block and/or Report other Users.
Visibility is now taken into effect when you attach a Discussion or Bulletin to a Group with Hidden Visibility and will only display content who are in that Group (which at this time is only you).

Also changed the create icon to a plus sign.

I'm working on Search functionality now (added a bit for it in this update).